+++ src/hooks/useMeasure.ts
@@ -1,0 +1,23 @@
+import React, { useRef, useState } from "react"
+import ResizeObserver from "resize-observer-polyfill"
+
+/**
+ * useMeasure
+ */
+export function useMeasure() {
+  const ref = useRef()
+
+  const [bounds, set] = useState({ left: 0, top: 0, width: 0, height: 0 })
+
+  const [ro] = useState(
+    () => new ResizeObserver(([entry]) => set(entry.contentRect))
+  )
+
+  // ro.observe on mount, and clean up
+  useEffect(() => {
+    if (ref.current) ro.observe(ref.current)
+    return () => ro.disconnect()
+  }, [])
+
+  return [{ ref }, bounds]
+}
